chore: types refactor
