+++ src/adapters/FloorAdapter.ts
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
-import type { Floor } from "../types";
+import type { Floor } from "../types/cartography";
 
 /**
  * Adapts the server response to our common Floor object,
