+++ src/adapters/UserAdapter.ts
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
-import type { User, UserForm } from "../types";
+import type { User, UserForm } from "../types/auth";
 
 export type UserServer = UserForm & { isManager: boolean; isStaff: boolean };
