+++ src/domains/realtime.ts
@@ -7,7 +7,8 @@
  */
 import { realtimePositionsMapper } from "../adapters/RealtimeAdapter";
 import type ApiBase from "../apiBase";
-import type { RealtimePositions, UUID } from "../types";
+import type { UUID } from "../types/models";
+import type { RealtimePositions } from "../types/realtime";
 
 export type ResponseRealtimePosition = {
   features: Record<string, unknown>[];
