+++ src/types/coordinates.ts
@@ -1,0 +1,11 @@
+type LatLng = {
+  lat: number;
+  lng: number;
+};
+
+type Cartesians = {
+  x: number;
+  y: number;
+};
+
+export type { LatLng, Cartesians };
